[MOVE] News 코드 파일 구분 #64
