+++ DefaultComponents-Accessibility/DefaultComponents-Accessibility/List/NewListViewControllerWithAccessibility/NewsListViewControllerWithAccessibility.swift
@@ -14,7 +14,7 @@ final class NewsListViewControllerWithAccessibility: DefaultViewController {
     var dataSource: DataSource!
     var snapshot: Snapshot!
     
-    private lazy var tableView = UITableView()
+    lazy var tableView = UITableView()
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -46,66 +46,5 @@ private extension NewsListViewControllerWithAccessibility {
     
     func configureConstraints() {
         view.addPinnedSubview(tableView, height: nil)
-        
-    }
-}
-
-extension NewsListViewControllerWithAccessibility: ButtonSupplementaryViewDelegate {
-    func buttonSupplementaryView(didTapButton button: UIButton) {
-        if pagable.total == News.samples.count {
-            let alert = UIAlertController(title: "마지막 페이지입니다", message: "모든 페이지를 확인하셨습니다.", preferredStyle: .alert)
-            alert.addAction(UIAlertAction(title: "확인", style: .default))
-            self.present(alert, animated: true)
-            return
-        }
-        
-        let startIndex = pagable.lastIndex + 1
-        let expectedEndIndex = pagable.lastIndex + pagable.display
-        
-        let endIndex = expectedEndIndex < News.samples.count ? expectedEndIndex : pagable.lastIndex + (News.samples.count - pagable.lastIndex - 1)
-        
-        let items = Array(News.samples[startIndex...endIndex])
-        updateSnapshot(newItems: items, after: pagable.lastIndex)
-        
-        if UIAccessibility.isVoiceOverRunning {
-            Task {
-                try? await announceNumOfAddedNews(count: endIndex - startIndex + 1)
-            }
-        }
-    }
-    
-    @MainActor
-    func announceNumOfAddedNews(count: Int) async throws {
-        try await Task.sleep(for: .seconds(0.8))
-        UIAccessibility.post(notification: .announcement, argument: "\(count)개의 뉴스가 추가되었습니다")
-    }
-}
-
-extension NewsListViewControllerWithAccessibility {
-    typealias DataSource = UITableViewDiffableDataSource<Int, News>
-    typealias Snapshot = NSDiffableDataSourceSnapshot<Int, News>
-    
-    func initializeSnapshot() {
-        pagable = Pagable(total: news.count, start: 0, display: 3, lastIndex: news.count - 1)
-        snapshot = Snapshot()
-        snapshot.appendSections([0])
-        snapshot.appendItems(news)
-        dataSource.apply(snapshot)
-    }
-    
-    func updateSnapshot(newItems items: [News], after index: Int) {
-        pagable.total = pagable.total + items.count
-        pagable.start = news.count
-        pagable.lastIndex = pagable.total - 1
-        snapshot.insertItems(items, afterItem: news[index])
-        news.append(contentsOf: items)
-        
-        dataSource.apply(snapshot)
-        
-        let indexPath = IndexPath(item: index + 1, section: 0)
-        guard let cell = tableView.cellForRow(at: indexPath) else { return }
-        Task {
-            try? await UIAccessibility.setFocus(to: cell, for: .seconds(2))
-        }
     }
 }
